feat(expense-modal): confirm when removing an expense

+++ src/components/data/buttonRenderers/RemoveButtonRenderer.tsx
@@ -1,5 +1,5 @@
 import { DeleteOutlined } from "@ant-design/icons";
-import { Button } from "antd";
+import { Button, Modal } from "antd";
 import React from "react";
 import expenseStore from "~/stores/expenseStore";
 
@@ -11,9 +11,13 @@ interface Props {
 // eslint-disable-next-line mobx/missing-observer
 const RemoveButtonRenderer: React.FC<Props> = ({ id, onClick }) => {
   const handleClick = React.useCallback(() => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    expenseStore.delete(id);
-    onClick();
+    Modal.confirm({
+      content: "Вы уверены, что хотите удалить этот расход?",
+      onOk: async () => {
+        await expenseStore.delete(id);
+        onClick();
+      },
+    });
   }, [id, onClick]);
 
   return <Button danger icon={<DeleteOutlined />} onClick={handleClick} />;
